refactored form component into its own file
