+++ src/component/contact-form.module.css
@@ -1,0 +1,47 @@
+.submit{
+    border: 2px solid #24a0ed; /* Light Blue */
+    background-color: #24a0ed;
+    border-radius: 10px;
+    color: black;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 20px;
+    margin-top: 20px;
+    padding: 15px 15px;
+    transition-duration: 0.4s;
+}
+.submit:hover{
+    border: 2px solid #0000FF;
+    background-color: #0000FF;
+    color: white;
+}
+.cancel{
+    border-radius: 10px;
+    color: black;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 20px;
+    margin-left: 15px;
+    margin-top: 20px;
+    padding: 15px 15px;
+    transition-duration: 0.4s;
+}
+.cancel:hover{
+    background-color: #FF0000;
+    color: black;
+}
+
+.inputName{
+    font-size: 18px;
+    float: left;
+    margin-left: 20px;
+}
+
+.inputField{
+    type: text;
+    border-weight: 5px;
+    width: 80%;
+    padding: 8px 14px;
+    margin: 8px 0;
+    font-size: 18px;
+}
