+++ src/component/contact-info.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import ContactForm from './contact-form.js';
 import styles from './contact-info.module.css';
 import axios from "axios";
 
@@ -11,7 +12,7 @@ const contactInfo = {
 }
 
 
-function ContactInfo() {
+export default function ContactInfo() {
 
   const [isEditing, setIsEditing] = useState(false);
 
@@ -19,43 +20,12 @@ function ContactInfo() {
     setIsEditing(!isEditing);
   };
 
-  const [contact, setContact] = useState({
-    firstName: contactInfo.firstName,
-    lastName: contactInfo.lastName,
-    phone: contactInfo.phone,
-    email: contactInfo.email,
-  });
-
-  function handleChange(event){
-    const { name, value } = event.target;
-    if(name === "firstName"){
-        setContact({
-            ...contact,
-            firstName: value,
-        });
+  // This function might need to be refactored.
+  async function AddContact(contact, didSubmit){
+    if(!didSubmit){
+        toggleEdit();
+        return;
     }
-    else if(name === "lastName"){
-        setContact({
-            ...contact,
-            lastName: value,
-        });
-    }
-    else if(name === "phone"){
-        setContact({
-            ...contact,
-            phone: value,
-        });
-    }
-    else{
-        setContact({
-           ...contact,
-           email: value,
-        });
-    }
-    return;
-  }
-
-  async function AddContact(){
     try {
         const response = await axios.post("http://localhost:8000/contact", contact);
         toggleEdit();
@@ -68,18 +38,8 @@ function ContactInfo() {
     }
   }
 
-  async function EditContact(){
-    try{
-        const initResponse = await axios.get("http://localhost:8000/contact", contact);
-        const response = axios.put('http://localhost:8000/contact/${initResponse._id}', contact);
-        toggleEdit();
-        return response;
-    } catch (error) {
-        console.log(error);
-        // Not handling errors at the moment
-        toggleEdit();
-        return false;
-    }
+  function EditContact(contact, didSubmit){
+    return;
   }
 
   function AddToGroup(){
@@ -101,54 +61,22 @@ function ContactInfo() {
       />
       <div className={styles.body}>
         {isEditing ?
-        // Form object. No real functionality yet.
-        <>
-        <div className={styles.inputName}> First Name... </div>
-        <input className={styles.inputField} placeholder="first name" name="firstName"
-               id="firstName" value={contact.firstName} onChange={handleChange}/>
-
-        <div className={styles.inputName}> Last Name... </div>
-        <input className={styles.inputField} placeholder="last name" name="lastName"
-               id="lastName" value={contact.lastName} onChange={handleChange}/>
-
-        <div className={styles.inputName}> Phone Number... </div>
-        <input className={styles.inputField} placeholder="phone number" name="phone"
-               id="phone" value={contact.phone} onChange={handleChange}/>
-
-        <div className={styles.inputName}> Email... </div>
-        <input className={styles.inputField} placeholder="email" name="email"
-               id="email" value={contact.email} onChange={handleChange}/>
-        </>
+        // Edit contact form
+        <ContactForm handleSubmit={AddContact} contact={contactInfo}/>
         :
         // Display
         <>
-        <h2 className={styles.name} > {contactInfo.firstName + ' ' + contactInfo.lastName} </h2>
-        <p className={styles.info}> {'Email: ' +  contactInfo.email } </p>
-        <p className={styles.info}> {'Phone: ' + contactInfo.phone} </p>
-        </>
-        }
-      </div>
-
-
-      <div>
-        {isEditing ?
-        <>
-        <input className={styles.submit} type = "submit" value = "Submit" onClick={AddContact} />
-        <input className={styles.cancel} type = "button" value = "Cancel" onClick={toggleEdit} />
-        </>
-        :
-        <>
-        <input className={styles.group} type = "button" value = "Add to Group" onClick={AddToGroup} />
-        <div>
-        <input className={styles.edit} type = "button" value = "Edit" onClick={toggleEdit} />
-        <input className={styles.deleteButton} type = "button" value = "Delete" onClick={DeleteContact} />
-        </div>
+            <h2 className={styles.name} > {contactInfo.firstName + ' ' + contactInfo.lastName} </h2>
+            <p className={styles.info}> {'Email: ' +  contactInfo.email } </p>
+            <p className={styles.info}> {'Phone: ' + contactInfo.phone} </p>
+            <input className={styles.group} type = "button" value = "Add to Group" onClick={AddToGroup} />
+            <div>
+                <input className={styles.edit} type = "button" value = "Edit" onClick={toggleEdit} />
+                <input className={styles.deleteButton} type = "button" value = "Delete" onClick={DeleteContact} />
+            </div>
         </>
         }
       </div>
     </div>
   );
 }
-
-
-export default ContactInfo;
